app/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
flask==3.0.0
APScheduler==3.10.4
pytz==2023.3
